Adds Templates For Data-Table-Posts and gatsby-node.js
I used the steps from the tutorial so that we can create pages out of MD files. I was thinking we could treat each data post like a blog post. It could be nice to display the Title of the post and date somewhere on the home page which links to the post itself. I also still want to look into how to create pages from MDX files. Might be nice to have both options built in.