Adds MDXProvider to Allow Use Of JSX Components.
I also had to delete the .cache in the directory because it was causing a really annoying error that would cause webpack to crash on build. This was all do to Gatsby looking for the components within the content directory I guess from the way it caches files? Either way it works now haha.